requirements.txt
Flask
Flask-SQLAlchemy
Pillow
scikit-learn
numpy
opencv-python-headless  
torch
torchvision
Werkzeug>=2.3.0 